fix: add missing #include <SD.h> in gps_track.h, fix narrowing warning
- gps_track.h: add #include <SD.h> so File type is declared
- display_manager.cpp: explicit uint16_t cast to fix narrowing warning

+++ src/display_manager.cpp
@@ -1188,8 +1188,8 @@ void DisplayManager::drawSetupScreen(int cursor, MachineMode mode,
     // Kolor wartosci
     uint16_t valColors[3] = {
         COLOR_ACCENT,
-        smartSwitch ? COLOR_ACCENT : COLOR_WARNING,
-        gapStart    ? COLOR_WARNING : COLOR_ACCENT
+        (uint16_t)(smartSwitch ? COLOR_ACCENT : COLOR_WARNING),
+        (uint16_t)(gapStart    ? COLOR_WARNING : COLOR_ACCENT)
     };
 
     const int SETUP_ITEM_H = 48;
